Add files via upload
This is a tricky simulation. So please approach me in class if  you have questions.
How does the ANOVA work? Can we re-create the ANOVA using sampling distributions? 
Often formulas are simply thrown at students, and many will nod politely. The truth is, that many researchers don't actually understand how this works.

The aim of this lesson is to help you gain an intuition of ANOVA using simulations.