Add files via upload
In this simulation we prove that two-way ANOVA does not inflate type I error rate when we have DV ~ IV1 + IV2. However, two one-way ANOVAs do. The Two-Way ANOVA holds the overall alpha (type I) at 0.05 because it uses a single Mean Square Error term calculated from all the data. I won't go into the maths for this one, but feel free to do some further reading if you would like to know how the F-statistics are derived for each group, as well as their interaction (this part is complex).